# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/config": {
      "post": {
        "operationId": "ConfigController_create",
        "summary": "Create a device configuration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The caller is signed in but lacks the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ConfigController_findAll",
        "summary": "List device configurations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Every configuration, unpaginated (discrepancy D6).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConfigResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/config/{id}": {
      "put": {
        "operationId": "ConfigController_update",
        "summary": "Update a device configuration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated configuration, or an empty body when no record has this id (discrepancy D2).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The caller is signed in but lacks the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ConfigController_updateAnalogLecure",
        "summary": "Report a tank level from the device (needs a sign-in and the device key in the body)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TankLevelConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated configuration, or an empty body when no record has this id (discrepancy D2). A wrong device key is refused with 404 (discrepancy D10).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No record exists with the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Config"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/activity": {
      "post": {
        "operationId": "ActivityController_create",
        "summary": "Create an activity",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created activity, with its type as an id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityRecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The caller is signed in but lacks the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ActivityController_findAll",
        "summary": "List activities",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Activity type id.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "COMPLETED",
                "TODO"
              ],
              "type": "string"
            }
          },
          {
            "name": "price",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching activities, newest first, unpaginated (discrepancy D6).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/activity/{id}": {
      "get": {
        "operationId": "ActivityController_findOne",
        "summary": "Get an activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The activity, with its type populated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No record exists with the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ActivityController_update",
        "summary": "Update an activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated activity with its type as an id, or an empty body when no record has this id (discrepancy D2).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityRecordResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The request body, query or path value is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ActivityController_remove",
        "summary": "Delete an activity",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The activity was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityDeletedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid credentials were supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The caller is signed in but lacks the required role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No record exists with the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Activity"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Checklist API",
    "description": "Documentación Checklist API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "User or password incorrect."
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateConfigDto": {
        "type": "object",
        "properties": {
          "doorLock": {
            "type": "string"
          },
          "mainLock": {
            "type": "string"
          },
          "usersLimit": {
            "type": "number"
          },
          "analogLecture": {
            "type": "number"
          }
        },
        "required": [
          "doorLock",
          "mainLock",
          "usersLimit",
          "analogLecture"
        ]
      },
      "ConfigResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "doorLock": {
            "type": "string"
          },
          "mainLock": {
            "type": "string"
          },
          "usersLimit": {
            "type": "number"
          },
          "analogLecture": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "doorLock",
          "mainLock",
          "usersLimit",
          "analogLecture",
          "createdAt",
          "updatedAt",
          "__v"
        ]
      },
      "UpdateConfigDto": {
        "type": "object",
        "properties": {
          "doorLock": {
            "type": "string"
          },
          "mainLock": {
            "type": "string"
          },
          "usersLimit": {
            "type": "number"
          },
          "analogLecture": {
            "type": "number"
          }
        }
      },
      "TankLevelConfigDto": {
        "type": "object",
        "properties": {
          "analogLecture": {
            "type": "number"
          },
          "apiKey": {
            "type": "string"
          },
          "time": {
            "type": "number"
          }
        },
        "required": [
          "analogLecture",
          "apiKey",
          "time"
        ]
      },
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "TODO"
            ],
            "type": "string",
            "default": "TODO"
          },
          "price": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "price",
          "date"
        ]
      },
      "ActivityRecordResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "TODO"
            ],
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "type",
          "status",
          "price",
          "date",
          "createdAt",
          "updatedAt",
          "__v"
        ]
      },
      "ActivityTypeResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "budget": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "name",
          "budget",
          "createdAt",
          "updatedAt",
          "__v"
        ]
      },
      "ActivityResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ActivityTypeResponseDto"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "TODO"
            ],
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "__v": {
            "type": "number"
          }
        },
        "required": [
          "_id",
          "type",
          "status",
          "price",
          "date",
          "createdAt",
          "updatedAt",
          "__v"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "status": {
            "enum": [
              "COMPLETED",
              "TODO"
            ],
            "type": "string",
            "default": "TODO"
          },
          "price": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ActivityDeletedResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}